Adding updates

+++ Inventory.Application/DataTransferObjects/LocationDto/LocationRequest.cs
@@ -4,7 +4,7 @@
     {
         public string Address { get; set; } = string.Empty;
         public string City { get; set; } = string.Empty;
-        public int CoordinateX { get; set; }
-        public int CoordinateY { get; set; }
+        public double CoordinateX { get; set; }
+        public double CoordinateY { get; set; }
     }
 }
